requirements.txt
# This file may be used to create an environment using:

tensorflow
keras
opencv-python
numpy
keras-video-generators
